refactor(loginscreen, capslockindicator): Caps Lock indicator class was made independent
