+++ src/widget/loginscreen.cpp
@@ -24,9 +24,6 @@
 #include "src/persistence/profilelocker.h"
 #include "src/nexus.h"
 #include "src/persistence/settings.h"
-#ifdef QTOX_PLATFORM_EXT
-#include "src/platform/capslock.h"
-#endif
 #include "src/widget/form/setpassworddialog.h"
 #include "src/widget/translator.h"
 #include "src/widget/style.h"
@@ -69,44 +66,6 @@ LoginScreen::LoginScreen(QWidget *parent) :
 
     retranslateUi();
     Translator::registerHandler(std::bind(&LoginScreen::retranslateUi, this), this);
-}
-
-LoginScreen::CapsLockIndicator::CapsLockIndicator(QWidget *parent) : QToolButton(parent) {
-    inputSize = QSize(130, 23);
-    cleanInputStyle = parentWidget()->styleSheet();
-
-    QIcon icon = QIcon(":img/caps_lock.svg");
-    setIcon(icon);
-    QSize iconSize(inputSize.height(), inputSize.height());
-    setIconSize(iconSize);
-    setCursor(Qt::ArrowCursor);
-    move(inputSize.width() - inputSize.height(), 0);
-    setStyleSheet("border: none; padding: 0;");
-    setToolTip(tr("CAPS-LOCK ENABLED"));
-}
-
-void LoginScreen::CapsLockIndicator::show() {
-    QToolButton::show();
-
-    QString style = QString("padding: -3px %1px -3px -6px").arg(iconSize().width() - 3);
-    parentWidget()->setStyleSheet(style);
-}
-
-void LoginScreen::CapsLockIndicator::hide() {
-    QToolButton::hide();
-    parentWidget()->setStyleSheet(cleanInputStyle);
-}
-
-void LoginScreen::CapsLockIndicator::updateIndicator() {
-    bool caps = false;
-    // It doesn't needed for OSX, because it shows indicator by default
-#ifdef QTOX_PLATFORM_EXT && !defined(Q_OS_OSX)
-    caps = Platform::capsLockEnabled();
-#endif
-    if (caps)
-        show();
-    else
-        hide();
 }
 
 LoginScreen::~LoginScreen()
